debugged

+++ lm_eval/models/custom_tree_srm.py
@@ -1416,7 +1416,7 @@ class SRMHFLM(TemplateLM):
             # tree selection only
             if not self.tree_expansion:
                 # cont = self.model.generate(context_enc, max_new_tokens=256, do_sample=True, top_p=0.9, temperature=0.7)
-                cont = torch.ones(50, 1024)
+                cont = torch.ones((50, 1024), dtype=torch.long)
                 with torch.no_grad():
                     # full_prompt = torch.cat((context_enc, cont), dim=1)
                     # print (self.tok_decode(cont[0]))
